Move PSNewTypeDeclarationImpl to it's own package
