+++ src/main/java/org/purescript/features/PSFindUsageProvider.kt
@@ -3,13 +3,13 @@ package org.purescript.features
 import com.intellij.lang.findUsages.FindUsagesProvider
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiNamedElement
-import org.purescript.file.PSFile
 import org.purescript.psi.*
 import org.purescript.psi.classes.PSClassDeclaration
 import org.purescript.psi.classes.PSClassMember
 import org.purescript.psi.data.PSDataConstructor
 import org.purescript.psi.data.PSDataDeclaration
 import org.purescript.psi.imports.PSImportAlias
+import org.purescript.psi.newtype.PSNewTypeDeclarationImpl
 import org.purescript.psi.typesynonym.PSTypeSynonymDeclaration
 
 class PSFindUsageProvider : FindUsagesProvider {
